backend/requirements.txt
# iCapture System - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask==2.3.2
flask-cors==4.0.0

# Database
pymysql==1.1.0
sqlalchemy==2.0.20

# Computer Vision
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0

# YOLOv5 Dependencies
torch==2.0.1
torchvision==0.15.2
pyyaml==6.0.1
tqdm==4.66.1

# OCR
pytesseract==0.3.10

# Utilities
python-dotenv==1.0.0
psutil==5.9.5

# Development (optional)
pytest==7.4.2
pytest-cov==4.1.0
